+++ zcash_proofs/src/circuit/sprout/mod.rs
@@ -35,12 +35,12 @@ pub struct JoinSplit {
     pub vpub_new: Option<u64>,
     pub h_sig: Option<[u8; 32]>,
     pub phi: Option<[u8; 32]>,
-    pub inputs: Vec<JSInput>,
-    pub outputs: Vec<JSOutput>,
+    pub inputs: Vec<JsInput>,
+    pub outputs: Vec<JsOutput>,
     pub rt: Option<[u8; 32]>,
 }
 
-pub struct JSInput {
+pub struct JsInput {
     pub value: Option<u64>,
     pub a_sk: Option<SpendingKey>,
     pub rho: Option<UniqueRandomness>,
@@ -48,7 +48,7 @@ pub struct JSInput {
     pub auth_path: [Option<([u8; 32], bool)>; TREE_DEPTH],
 }
 
-pub struct JSOutput {
+pub struct JsOutput {
     pub value: Option<u64>,
     pub a_pk: Option<PayingKey>,
     pub r: Option<CommitmentRandomness>,
@@ -389,7 +389,7 @@ fn test_sprout_constraints() {
             let r = Some(CommitmentRandomness(get_u256(&mut test_vector)));
             let a_sk = Some(SpendingKey(get_u256(&mut test_vector)));
 
-            inputs.push(JSInput {
+            inputs.push(JsInput {
                 value,
                 a_sk,
                 rho,
@@ -406,7 +406,7 @@ fn test_sprout_constraints() {
             get_u256(&mut test_vector);
             let r = Some(CommitmentRandomness(get_u256(&mut test_vector)));
 
-            outputs.push(JSOutput { value, a_pk, r });
+            outputs.push(JsOutput { value, a_pk, r });
         }
 
         let vpub_old = Some(test_vector.read_u64::<LittleEndian>().unwrap());