+++ zcash_client_sqlite/CHANGELOG.md
@@ -8,6 +8,12 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 ### Changed
 - MSRV is now 1.51.0.
+- Renamed the following to use lower-case abbreviations (matching Rust
+  naming conventions):
+  - `zcash_client_sqlite::BlockDB` to `BlockDb`
+  - `zcash_client_sqlite::WalletDB` to `WalletDb`
+  - `zcash_client_sqlite::error::SqliteClientError::IncorrectHRPExtFVK` to
+    `IncorrectHrpExtFvk`.
 
 ## [0.3.0] - 2021-03-26
 This release contains a major refactor of the APIs to leverage the new Data